requirements.txt
ldap3
django
djangorestframework
